Consolidated new review and update review into same form component
