+++ react-app/src/App.js
@@ -12,7 +12,6 @@ import UserProfilePage from "./components/UserProfilePage";
 import BusinessForm from "./components/Businesses/BusinessForm";
 import UpdateBusiness from "./components/Businesses/UpdateBusiness"
 import ReviewForm from "./components/ReviewForm";
-import NewReviewForm from "./components/ReviewForm/NewReview";
 
 function App() {
   const dispatch = useDispatch();
@@ -45,7 +44,7 @@ function App() {
             <UpdateBusiness />
           </Route>
           <Route exact path="/business/:id/review">
-            <NewReviewForm />
+            <ReviewForm />
           </Route>
           <Route exact path="/business/:id">
             <BusinessDetails />
@@ -54,7 +53,7 @@ function App() {
             <UserProfilePage />
           </Route>
           <Route exact path="/review/:reviewId">
-            <ReviewForm />
+            <ReviewForm isUpdate={true} />
           </Route>
         </Switch>
       )}
